+++ src/newsletter-editor/utils.js
@@ -1,7 +1,17 @@
 /**
+ * WordPress dependencies
+ */
+import apiFetch from '@wordpress/api-fetch';
+
+/**
  * Internal dependencies
  */
 import { getServiceProvider } from '../service-providers';
+
+/**
+ * External dependencies
+ */
+import mjml2html from 'mjml-browser';
 
 export const getEditPostPayload = newsletterData => {
 	return {
@@ -26,3 +36,27 @@ export const validateNewsletter = newsletterData => {
  * @return {boolean} True if it contains a valid email string.
  */
 export const hasValidEmail = string => /\S+@\S+/.test( string );
+
+/**
+ * Refresh the email-compliant HTML for a post.
+ *
+ * @param {number} postId      The current post ID.
+ * @param {string} postTitle   The current post title.
+ * @param {string} postContent The current post content.
+ * @return {Promise<string>} The refreshed email HTML.
+ */
+export const refreshEmailHtml = async ( postId, postTitle, postContent ) => {
+	const mjml = await apiFetch( {
+		path: `/newspack-newsletters/v1/post-mjml`,
+		method: 'POST',
+		data: {
+			post_id: postId,
+			title: postTitle,
+			content: postContent,
+		},
+	} );
+
+	// Once received MJML markup, convert it to email-compliant HTML and save as post meta.
+	const { html } = mjml2html( mjml, { keepComments: false, minify: true } );
+	return html;
+};